+++ sample-basic-java/src/main/java/org/dbtools/sample/model/database/main/individual/Individual.java
@@ -8,27 +8,6 @@
 
 package org.dbtools.sample.model.database.main.individual;
 
-import android.database.Cursor;
-import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues;
-
-
 public class Individual extends IndividualBaseRecord {
-
-
-    public Individual(Individual record) {
-        super(record);
-    }
-
-    public Individual(Cursor cursor) {
-        setContent(cursor);
-    }
-
-    public Individual(DBToolsContentValues values) {
-        setContent(values);
-    }
-
-    public Individual() {
-    }
-
 
 }