+++ sample-basic-java/src/main/java/org/dbtools/sample/model/database/main/individualquery/IndividualQuery.java
@@ -8,9 +8,6 @@
 
 package org.dbtools.sample.model.database.main.individualquery;
 
-import android.database.Cursor;
-
-import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues;
 import org.dbtools.sample.model.database.main.individual.IndividualConst;
 
 public class IndividualQuery extends IndividualQueryBaseRecord {
@@ -23,20 +20,6 @@ public class IndividualQuery extends IndividualQueryBaseRecord {
             ")";
     public static final String QUERY_RAW = "SELECT * FROM " + QUERY;
 
-    public IndividualQuery(IndividualQuery record) {
-        super(record);
-    }
-
-    public IndividualQuery(Cursor cursor) {
-        setContent(cursor);
-    }
-
-    public IndividualQuery(DBToolsContentValues values) {
-        setContent(values);
-    }
-
-    public IndividualQuery() {
-    }
 
 
 }