+++ sample-basic-java/src/main/java/org/dbtools/sample/model/database/main/individualview/IndividualViewBaseRecord.java
@@ -22,11 +22,6 @@ public abstract class IndividualViewBaseRecord extends AndroidBaseRecord {
     private Long id = null;
     private String name = "";
 
-    public IndividualViewBaseRecord(IndividualView record) {
-        this.id = record.getId();
-        this.name = record.getName();
-    }
-
     public IndividualViewBaseRecord() {
     }