+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/DatabaseBaseManager.kt
@@ -13,7 +13,6 @@ package org.dbtools.sample.kotlin.model.database
 import org.dbtools.android.domain.AndroidDatabase
 import org.dbtools.android.domain.AndroidBaseManager
 import org.dbtools.android.domain.AndroidDatabaseManager
-import org.dbtools.android.domain.database.DatabaseWrapper
 import org.dbtools.android.domain.config.DatabaseConfig