+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/main/individual/Individual.kt
@@ -1,32 +1,16 @@
 /*
  * Individual.kt
  *
- * Created: 05/09/2016 09:21:13
+ * Created: 01/28/2017 11:17:12
  */
 
 
 
 package org.dbtools.sample.kotlin.model.database.main.individual
 
-import android.database.Cursor
-import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues
 
+class Individual : IndividualBaseRecord() {
 
-class Individual : IndividualBaseRecord {
-
-    constructor(record: Individual) : super(record) {
-    }
-
-    constructor() {
-    }
-
-    constructor(cursor: Cursor) {
-        setContent(cursor)
-    }
-
-    constructor(values: DBToolsContentValues<*>) {
-        setContent(values)
-    }
 
 
 }