+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/main/individualdata/IndividualData.kt
@@ -8,26 +8,8 @@
 
 package org.dbtools.sample.kotlin.model.database.main.individualdata
 
-import android.database.Cursor
-import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues
 
-
-class IndividualData : IndividualDataBaseRecord {
-
-
-    constructor() {
-    }
-
-    constructor(record: IndividualData) : super(record) {
-    }
-
-    constructor(cursor: Cursor) {
-        setContent(cursor)
-    }
-
-    constructor(values: DBToolsContentValues<*>) {
-        setContent(values)
-    }
+class IndividualData : IndividualDataBaseRecord() {
 
 
 }