+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/main/individualdata/IndividualDataBaseRecord.kt
@@ -16,19 +16,13 @@ import org.dbtools.android.domain.database.contentvalues.DBToolsContentValues
 import android.database.Cursor
 
 
-@Suppress("LeakingThis", "unused", "RemoveEmptySecondaryConstructorBody")
+@Suppress("LeakingThis", "unused", "RemoveEmptySecondaryConstructorBody", "ConvertSecondaryConstructorToPrimary")
 @SuppressWarnings("all")
 abstract class IndividualDataBaseRecord : AndroidBaseRecord {
 
      open var externalId: Long? = 0
      open var typeId: Int? = 0
      open var name: String? = ""
-
-    constructor(record: IndividualData) {
-        this.externalId = record.externalId
-        this.typeId = record.typeId
-        this.name = record.name
-    }
 
     constructor() {
     }
@@ -62,6 +56,7 @@ abstract class IndividualDataBaseRecord : AndroidBaseRecord {
         return copy
     }
 
+    @Suppress("UNNECESSARY_NOT_NULL_ASSERTION")
     override fun bindInsertStatement(statement: StatementWrapper) {
         if (externalId != null) {
             statement.bindLong(1, externalId!!)
@@ -80,6 +75,7 @@ abstract class IndividualDataBaseRecord : AndroidBaseRecord {
         }
     }
 
+    @Suppress("UNNECESSARY_NOT_NULL_ASSERTION")
     override fun bindUpdateStatement(statement: StatementWrapper) {
         if (externalId != null) {
             statement.bindLong(1, externalId!!)