+++ sample-basic-kotlin/src/main/kotlin/org/dbtools/sample/kotlin/model/database/main/individualtype/IndividualType.kt
@@ -11,8 +11,6 @@
 package org.dbtools.sample.kotlin.model.database.main.individualtype
 
 
-@Suppress("LeakingThis", "unused", "RemoveEmptySecondaryConstructorBody")
-@SuppressWarnings("all")
 enum class IndividualType {
 HEAD, SPOUSE, CHILD;