+++ sample-rxjava/src/main/java/org/dbtools/sample/model/database/main/individualdata/IndividualDataBaseRecord.java
@@ -23,12 +23,6 @@ public abstract class IndividualDataBaseRecord extends AndroidBaseRecord {
     private Integer typeId = null;
     private String name = "";
 
-    public IndividualDataBaseRecord(IndividualData record) {
-        this.externalId = record.getExternalId();
-        this.typeId = record.getTypeId();
-        this.name = record.getName();
-    }
-
     public IndividualDataBaseRecord() {
     }