+++ sample-rxjava/src/main/java/org/dbtools/sample/model/database/main/individualquery/IndividualQueryBaseRecord.java
@@ -22,11 +22,6 @@ public abstract class IndividualQueryBaseRecord extends AndroidBaseRecord {
     private Long id = null;
     private String name = "";
 
-    public IndividualQueryBaseRecord(IndividualQuery record) {
-        this.id = record.getId();
-        this.name = record.getName();
-    }
-
     public IndividualQueryBaseRecord() {
     }